@@ -9,6 +9,8 @@
 
 namespace FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
 
+use Symfony\Component\Form\Extension\Core\Type\DateType as DateFormType;
+
 class DateType extends AbstractType
 {
     /**
@@ -24,6 +26,6 @@ class DateType extends AbstractType
      */
     protected function getFormType()
     {
-        return 'date';
+        return DateFormType::class;
     }
 }
