@@ -9,6 +9,8 @@
 
 namespace FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
 
+use Symfony\Component\Form\Extension\Core\Type\TimeType as TimeFormType;
+
 class TimeType extends AbstractType
 {
     /**
@@ -24,6 +26,6 @@ class TimeType extends AbstractType
      */
     protected function getFormType()
     {
-        return 'time';
+        return TimeFormType::class;
     }
 }
